# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trip Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/packing:
    get:
      operationId: listPackingItems
      tags: [packing]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Packing items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackingItem"
    post:
      operationId: createPackingItem
      tags: [packing]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
  /trips/{tripId}/packing/{itemId}:
    patch:
      operationId: updatePackingItem
      tags: [packing]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
    delete:
      operationId: deletePackingItem
      tags: [packing]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/notes:
    get:
      operationId: listTripNotes
      tags: [notes]
      summary: Get private notes for current user on this trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripNote"
    post:
      operationId: createTripNote
      tags: [notes]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
  /trips/{tripId}/notes/{noteId}:
    patch:
      operationId: updateTripNote
      tags: [notes]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripNoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripNote"
    delete:
      operationId: deleteTripNote
      tags: [notes]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trips/{tripId}/documents:
    get:
      operationId: listTravelDocuments
      tags: [documents]
      summary: Get private travel documents for current user on this trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelDocument"
    post:
      operationId: createTravelDocument
      tags: [documents]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelDocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelDocument"
  /trips/{tripId}/documents/{documentId}:
    patch:
      operationId: updateTravelDocument
      tags: [documents]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TravelDocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelDocument"
    delete:
      operationId: deleteTravelDocument
      tags: [documents]
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    PackingItem:
      type: object
      required: [id, tripId, name, checked]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        name:
          type: string
        category:
          type: ["string", "null"]
        checked:
          type: boolean
        required:
          type: boolean
    PackingItemInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        category:
          type: string
        checked:
          type: boolean
        required:
          type: boolean
    PackingItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        checked:
          type: boolean
        required:
          type: boolean
    TripNote:
      type: object
      required: [id, tripId, userId, content, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TripNoteInput:
      type: object
      required: [content]
      properties:
        title:
          type: string
        content:
          type: string
    TripNoteUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    TravelDocument:
      type: object
      required: [id, tripId, userId, type, number]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [passport, visa, insurance, id_card, other]
        number:
          type: string
        expiryDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    TravelDocumentInput:
      type: object
      required: [type, number]
      properties:
        type:
          type: string
          enum: [passport, visa, insurance, id_card, other]
        number:
          type: string
        expiryDate:
          type: string
        notes:
          type: string
    TravelDocumentUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [passport, visa, insurance, id_card, other]
        number:
          type: string
        expiryDate:
          type: string
        notes:
          type: string